+++ server/src/graphql/resolvers/auth.ts
@@ -1,0 +1,39 @@
+import bcrypt from "bcryptjs";
+
+import User from "../../models/userModel";
+import { dateToString } from "../../helpers/date";
+
+export default {
+    createUser: async (args: any) => {
+        try {
+            const userExisting = await User.findOne({
+                email: args.userInput.email,
+            });
+            if (userExisting) throw new Error("User exist already.");
+            const hashedPassword = await bcrypt.hash(
+                args.userInput.password,
+                12
+            );
+            const user = new User({
+                firstName: args.userInput.firstName,
+                lastName: args.userInput.lastName,
+                email: args.userInput.email,
+                password: hashedPassword,
+                phone: args.userInput.phone,
+                bio: args.userInput.bio,
+                website: args.userInput.website,
+                isProvider: args.userInput.isProvider,
+                image_url: args.userInput.image_url,
+                createdAt: new Date().toISOString(),
+            });
+
+            const result = await user.save();
+            return {
+                ...result._doc,
+                createdAt: dateToString(result.createdAt),
+            };
+        } catch (err) {
+            throw err;
+        }
+    },
+};
